feat: refactor project into more modular components and search functionality working
