+++ components/Ui/Backdrop.vue
@@ -1,6 +1,5 @@
 <template>
   <div class="index-backdrop">
-    <h1>The backadsfasdfasfasdfsdfasdfasdfasdfsdfasdfasdf</h1>
     <slot></slot>
   </div>
 </template>
@@ -10,4 +9,5 @@ export default {
   name: "Backdrop",
 };
 </script>
+
