+++ core/api/unsplash.js
@@ -2,4 +2,7 @@ export default ($unsplashAxios) => ({
   GetPhotos() {
     return $unsplashAxios.get(`/photos`);
   },
+  SearchPhotos(searchParams) {
+    return $unsplashAxios.get(`/search/photos?query=${searchParams}`);
+  },
 });
